as a test, Getting Emotes available on a channel
This currently requires PuppeteerSharp, so first run is going to be slow as it needs to download chromium into the working directory.. Pretty overkill. But until I figure out their OAuth2 and how to register an App to get a ClientID/ClientSecret for proper integration then this is the only way around cloudflare, etc.